# docs/technical/03_api_specification.generated.yaml
components:
  schemas:
...
    DashboardSummary:
      properties:
        donors_total:
          title: Donors Total
          type: integer
        orphans_available:
          title: Orphans Available
          type: integer
        orphans_sponsored:
          title: Orphans Sponsored
          type: integer
        orphans_total:
          title: Orphans Total
          type: integer
        payments_last_30d_count:
          title: Payments Last 30D Count
          type: integer
        payments_last_30d_total:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Payments Last 30D Total
          type: string
        sponsorships_active:
          title: Sponsorships Active
          type: integer
        sponsorships_overdue:
          title: Sponsorships Overdue
          type: integer
      required:
      - orphans_total
      - orphans_sponsored
      - orphans_available
      - donors_total
      - sponsorships_active
      - sponsorships_overdue
      - payments_last_30d_total
      - payments_last_30d_count
      title: DashboardSummary
      type: object
...
    DonationsByPartner:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PartnerDonations'
          title: Items
          type: array
        window_days:
          title: Window Days
          type: integer
      required:
      - window_days
      - items
      title: DonationsByPartner
      type: object
...
    MonthlyPoint:
      properties:
        month:
          format: date
          title: Month
          type: string
        payments_count:
          title: Payments Count
          type: integer
        payments_total:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Payments Total
          type: string
      required:
      - month
      - payments_total
      - payments_count
      title: MonthlyPoint
      type: object
...
    PartnerDonations:
      properties:
        partner_code:
          title: Partner Code
          type: string
        partner_id:
          format: uuid
          title: Partner Id
          type: string
        partner_name:
          title: Partner Name
          type: string
        payments_count:
          title: Payments Count
          type: integer
        payments_total:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Payments Total
          type: string
      required:
      - partner_id
      - partner_code
      - partner_name
      - payments_total
      - payments_count
      title: PartnerDonations
      type: object
...
    PaymentsTimeseries:
      properties:
        months:
          items:
            $ref: '#/components/schemas/MonthlyPoint'
          title: Months
          type: array
      required:
      - months
      title: PaymentsTimeseries
      type: object
...
    SponsorshipsByStatus:
      properties:
        slices:
          items:
            $ref: '#/components/schemas/StatusSlice'
          title: Slices
          type: array
      required:
      - slices
      title: SponsorshipsByStatus
      type: object
    StatusSlice:
      properties:
        count:
          title: Count
          type: integer
        status:
          title: Status
          type: string
      required:
      - status
      - count
      title: StatusSlice
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  title: Rufaqaa API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/stats/donations-by-partner:
    get:
      description: 'Completed payments rolled up by the orphan''s partner

        organization over the last 90 days. Top 10 partners by total,

        descending. Payments not linked to an orphan (e.g. general

        donations) are skipped.'
      operationId: donations_by_partner_api_v1_stats_donations_by_partner_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationsByPartner'
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Donations By Partner
      tags:
      - stats
  /api/v1/stats/payments-timeseries:
    get:
      description: Total paid + count grouped by month for the last 12 months.
      operationId: payments_timeseries_api_v1_stats_payments_timeseries_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsTimeseries'
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Payments Timeseries
      tags:
      - stats
  /api/v1/stats/sponsorships-by-status:
    get:
      description: 'One row per status enum that currently has at least one

        sponsorship attached. The frontend renders this as a small donut.'
      operationId: sponsorships_by_status_api_v1_stats_sponsorships_by_status_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorshipsByStatus'
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Sponsorships By Status
      tags:
      - stats
  /api/v1/stats/summary:
    get:
      operationId: dashboard_summary_api_v1_stats_summary_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Dashboard Summary
      tags:
      - stats
...